Created AuthorRepository + CleanUp
Split CheepRepository in two. There is now a CheepRepository and an AuthorRepository. There has been a general clean-up in most classes
