+++ src/Chirp.Web/Pages/UserTimeline.cshtml
@@ -40,19 +40,20 @@
                         {
                             @if (Model.followedAuthors.Any(a => a.Name == cheep.Author))
                             {
-                            <form method="Post" asp-page-handler="Unfollow" asp-route-followAuthorName="@cheep.Author" style="display:inline;">
-                                <button type="submit"style="padding: 2px 5px; height: 18px; width: auto;">Unfollow</button>
-                            </form>
+                                <form method="Post" asp-page-handler="Unfollow" asp-route-followAuthorName="@cheep.Author" style="display:inline;">
+                                    <button type="submit"style="padding: 2px 5px; height: 18px; width: auto;">Unfollow</button>
+                                </form>
                             }
                             else
                             {
-                            <form method="Post" asp-page-handler="Follow" asp-route-followAuthorName="@cheep.Author" style="display:inline;">
-                                <button type="submit"style="padding: 2px 5px; height: 18px; width: auto;">Follow</button>
-                            </form>
+                                <form method="Post" asp-page-handler="Follow" asp-route-followAuthorName="@cheep.Author" style="display:inline;">
+                                    <button type="submit"style="padding: 2px 5px; height: 18px; width: auto;">Follow</button>
+                                </form>
                             }
                         }
                     </div>
                         @cheep.Text
+                        <br/>
                         <small>&mdash; @cheep.TimeStamp</small>
                     </p>
                 </li>
